Implement entity interpolation, move tickrate into game settings
Entity interpolation allows us to render objects at a position that doesn't directly match the position of an entity, but instead smooths out entity movement, scaling, and rotation based on values from the current and previous tick. For example, the player controller now feels smooth even at high framerates (144fps tested) even at tickrates that do not match (10 to 200 tps tested). Eventually logic will be done on another thread, so changing the timestep like this helps us decouple the engine logic and rendering.

Game settings can now define a target tick rate. The default (and recommended) is 60 tps, but the engine should be able to scale up higher depending on the number of entities and scene complexity.
SpaceGame uses a tickrate of 60.