Split `VulkanRenderContext` source & header into sections too
The source file is split up on a per-class basis because some classes are massive (notably `VulkanRenderContext` itself).
The header file is split up in a more generic way because it's only a few hundred lines long, and it's better to be able to see the contents of a header file at a glance.