Fix #13: submit ImGui dummy item to validate extent in `DebugOverlay`
This was caused by a breaking change in ImGui v1.89 - see here: https://github.com/ocornut/imgui/releases/tag/v1.89#:~:text=Instead%2C%20please-,submit%20an%20item,-%3A%0ABegin(...).
Since we're getting the position of the main viewport and then moving the cursor based on that, the cursor position becomes negative when moving to the left of the default Windows monitor (eg. on dual-monitor setups where the main monitor is the one on the right).

I don't think there should be any other places where this applies, since `DebugOverlay` is the only area where we're setting the position of an element based on the position of the main viewport (ie. the engine window).