Fix for #15: Remove `ReloadShaders` from `Material`
This often caused an issue where the following steps would occur:
1. A material is created. This invokes the asset compiler, causing a change on the filesystem.
2. A FileSystemWatcher is created that watches for all file changes in the shaders directory.
3. A `NativeMaterial` is created. This creates resources including descriptors, pipelines, and textures where necessary.
4. Meanwhile, on the managed thread, `NativeMaterial.ReloadShaders()` would be called by the FileSystemWatcher created in step 2. This would execute on the managed thread.

This would cause the descriptor pool to be used inside both the managed and native threads, which is not allowed (they are externally synchronized: https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VkDescriptorPool.html).
Descriptor pools could, in future, be handled in the same way as a `VulkanCommandContext` where one is lazy intialised for each thread. They could also just be created as part of a `VulkanCommandContext`.

`ReloadShaders` doesn't actually work at the moment so I've removed it entirely while I re-think the way asset hot-loading is handled.