Add a number of options (see desc)
This commit adds the following options to the project manifest:

languageVersion: Defines the "LangVersion" option in the csproj. Default is "latest"

allowUnsafeBlocks: Defines the "AllowUnsafeBlocks" option in the csproj. Default is false.

useMochaGlobal: Whether or not to statically use "Mocha.Common.Global" in the project. Default is true.

usings: An array of using entries. Each using entry contains the namespace and whether or not to statically use it.

packageReferences: An array of NuGet package references. Options supported are: name, version, privateAssets, and includeAssets.

projectReferences: An array of project references. Options supported are: path, privateAssets, referenceOutputAssembly, and outputItemType.

rawEntry: Appends the literal string to the csproj.