New shader file format (see desc)
Sections are defined with a name and then the contents in curly brackets {}
Current supported sections:
- Vertex
- Fragment
- Compute

Additionally, a "Common" section can be used to insert preamble before all sections.
The "Shader" header currently isn't hooked up to anything, but will be used in the near future, so it's a good idea to fill this section out ahead-of-time.

This commit also moves "BaseParser" into Mocha.Common, so that it can be used in MochaTool.AssetCompiler along with its existing usage in Mocha.UI.