Gigarefactor 4: Re-write all bindings (see desc)
We don't want bindings to be "global" like they are now. Ideally we want to make it so that C# can use bindings based on whichever context / realm / etc. it's in, and we want to be able to spin up multiple clients & servers if necessary. So in order to do that, we're going to do class-specific bindings rather than the monolithic namespaced bindings we have now. This should hopefully also make things a little bit easier to manage.

Note that the engine won't run properly at the minute because I haven't updated anything to use these new bindings - they're still a work-in-progress.

I also want to do a few things in order to make bindings more powerful, as well as make bindings automatically translate from C++ types to interop-compatible C types - that'll probably be what I tackle next.