Mocha::HandleMap refactor (see desc)
This refactors only the Framework version of the HandleMap, which isn't being used anywhere in the engine yet.
- Use std::vector as backing collection instead of std::unordered_map
- Remove HandleMap::Find(), HandleMap::RemoveAt()
- Speed things up significantly by returning const refs to unique_ptr instead of copying shared_ptrs
- Add operator[], Front(), and Back()

This provides a significant speedup - from ~7M lookups/s to ~16M/s. This still isn't as fast as using Mocha::Array (~181M/s) so there's still a lot of work that could be done here.